+++ game_patch/hud/hud_world.h
@@ -63,7 +63,8 @@ enum class WorldHUDRenderMode : int
 {
     no_overdraw,
     no_overdraw_glow,
-    overdraw
+    overdraw,
+    overdraw_colorized
 };
 
 struct EphemeralWorldHUDSprite
@@ -77,6 +78,7 @@ struct EphemeralWorldHUDSprite
     int duration = 10000;
     bool float_away = false;
     float wind_phase_offset = 0.0f;
+    rf::Color color{255, 255, 255, 255};
 };
 
 struct EphemeralWorldHUDString